compiling warning addressed

+++ include/fast_limo/Modules/Mapper.cpp
@@ -66,7 +66,7 @@
             init_matches.resize(pc->points.size()-N0);
 
             #pragma omp parallel for num_threads(this->num_threads_)
-            for(int i = 0; i < pc->points.size()-N0; i++){
+            for(size_t i= 0; i < pc->points.size()-N0; i++){
                 
                 Eigen::Vector4f bl4_point(pc->points[i].x, pc->points[i].y, pc->points[i].z, 1.); // base link 4d point
                 Eigen::Vector4f global_point = s.get_RT() * bl4_point;                            // global 4d point == [x', y', z', 1.0]
@@ -76,7 +76,7 @@
             }
 
             Matches chosen_matches;
-            for(int j = 0; j < init_matches.size(); j++){
+            for(size_t j = 0; j < init_matches.size(); j++){
                 if(init_matches[j].lisanAlGaib())
                     chosen_matches.push_back(init_matches[j]); // if match is chosen, push it
             }
